requirements.txt
#-- LangChain and related packages
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.9
langchain-chroma==0.1.4

#-- Vector database
chromadb==0.5.20

#-- LLMs
openai==1.57.4
httpx==0.27.2

#-- API server
fastapi==0.115.5
uvicorn[standard]==0.32.1

#-- PDF processing
pdfplumber==0.11.4
pypdf==5.1.0

# -- Utilities
python-dotenv==1.0.1
tiktoken==0.8.0

# -- Unused
#streamlit==1.40.1
#pandas==2.2.3

